requirements.txt
llama-index
google-genai
langchain-core 
langchain-google-vertexai
python-dotenv
python-pptx
requests